# docs/contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SuccessEnvelope_AgentArtifactView_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AgentArtifactView"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "const": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ],
        "title": "SuccessEnvelope[AgentArtifactView]",
        "type": "object"
      },
...
      "SuccessEnvelope_ContextSearchResult_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ContextSearchResult"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "const": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ],
        "title": "SuccessEnvelope[ContextSearchResult]",
        "type": "object"
      },
      "SuccessEnvelope_DeletedResource_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeletedResource"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "success": {
            "const": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ],
        "title": "SuccessEnvelope[DeletedResource]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Bioinfoflow API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agent/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_api_v1_agent_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_AgentArtifactView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/artifacts/{artifact_id}/download": {
      "get": {
        "operationId": "download_artifact_api_v1_agent_artifacts__artifact_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Artifact download bytes"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/attachments/{attachment_id}": {
      "delete": {
        "operationId": "delete_attachment_api_v1_agent_attachments__attachment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "title": "Attachment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_DeletedResource_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Attachment",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/attachments/{attachment_id}/preview": {
      "get": {
        "operationId": "preview_attachment_api_v1_agent_attachments__attachment_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "title": "Attachment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Attachment preview bytes"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Attachment",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/context/search": {
      "get": {
        "operationId": "search_context_api_v1_agent_context_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 500,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "mixed",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_ContextSearchResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Context",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/sessions": {
      "get": {
        "operationId": "list_sessions_api_v1_agent_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_list_AgentSessionSummary__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "agent"
        ]
      },
      "post": {
        "operationId": "create_session_api_v1_agent_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_SessionSnapshot_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/v1/agent/sessions/{session_id}/artifacts": {
      "get": {
        "operationId": "list_session_artifacts_api_v1_agent_sessions__session_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_list_AgentArtifactView__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Session Artifacts",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}